+++ src/main/java/me/drex/instantfeedback/config/ConfigManager.java
@@ -2,12 +2,15 @@ package me.drex.instantfeedback.config;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
 import net.fabricmc.loader.api.FabricLoader;
 
+import java.util.HashSet;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Set;
 
 import static me.drex.instantfeedback.InstantFeedback.LOGGER;
 
@@ -18,6 +21,8 @@ public class ConfigManager {
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
     private static Config config = new Config();
 
+    private static Set<String> enabledFeatures = new HashSet<>();
+
     public static boolean load() {
         LOGGER.info("Loading instantfeedback config");
         if (Files.exists(CONFIG_FILE)) {
@@ -26,6 +31,7 @@ public class ConfigManager {
                 try {
                     config = GSON.fromJson(data, Config.class);
                     Files.writeString(CONFIG_FILE, GSON.toJson(config));
+                    loadEnabledFeatures();
                     return true;
                 } catch (JsonSyntaxException e) {
                     LOGGER.error("Failed to parse instantfeedback config", e);
@@ -36,6 +42,7 @@ public class ConfigManager {
         } else {
             try {
                 Files.writeString(CONFIG_FILE, GSON.toJson(config));
+                loadEnabledFeatures();
                 return true;
             } catch (IOException e) {
                 LOGGER.error("Failed to save instantfeedback config", e);
@@ -44,8 +51,23 @@ public class ConfigManager {
         return false;
     }
 
+    private static void loadEnabledFeatures() {
+        enabledFeatures.clear();
+        JsonObject jsonObject = GSON.toJsonTree(config).getAsJsonObject();
+        jsonObject.keySet().forEach(key -> {
+            if (jsonObject.get(key).isJsonPrimitive() && jsonObject.get(key).getAsJsonPrimitive().isBoolean()) {
+                if (jsonObject.get(key).getAsBoolean()) {
+                    enabledFeatures.add(key);
+                }
+            }
+        });
+    }
+
     public static Config config() {
         return config;
     }
 
+    public static Set<String> enabledFeatures() {
+        return enabledFeatures;
+    }
 }